# Improve `get_block_templates` performance