Trac ticket https://core.trac.wordpress.org/ticket/57756

## What

This PR improves the performance of block template resolution.

## Why

It provides a performance improvement for block themes. This change should not affect themes that don't have block templates, given they [bail early](https://github.com/WordPress/wordpress-develop/blob/4049c5d3543e59f92d48fd9cef735420996dce2e/src/wp-includes/block-template.php#L51).

The effect of this PR according to the [performance CI job](https://github.com/WordPress/wordpress-develop/actions/runs/4733710372/jobs/8401618173?pr=4097):

- Home classic theme

| Metric (ms) | This PR | Base | Improvement |
| --- | --- | --- | --- |
| wpBeforeTemplate | 17.02 | 19.47 | +12.58% |
|    wpTemplate    | 41.06 |  44.40 | +7.52% |
|     wpTotal      | 57.92 | 64.23 | +9.82% |

- Home block theme

| Metric (ms) | This PR | Base | Improvement % |
| --- | --- | --- | --- |
| wpBeforeTemplate | 23.34 | 24.42 | +4.42% |
|    wpTemplate    | 44.27 | 55.24 | +19.85% |
|     wpTotal      | 67.63 | 79.56 | +14.99% |


## How

The `get_block_templates` function is responsible to find block templates that match a given search. The function is provided a query parameter that is used to find the relevant user templates (database) and theme templates (file directory). The query parameter includes data such the slugs of the templates, the areas of the template parts, etc.

I've found that the `get_block_templates` retrieves and processes [all block templates provided by the theme](https://github.com/WordPress/wordpress-develop/blob/4049c5d3543e59f92d48fd9cef735420996dce2e/src/wp-includes/block-template-utils.php#L977), to [filter out the ones that don't match](https://github.com/WordPress/wordpress-develop/blob/4049c5d3543e59f92d48fd9cef735420996dce2e/src/wp-includes/block-template-utils.php#L997) after. This function can be more performant if it filters out the ones that don't match _before_ processing them, so it only computes the data to be used.

The following profile shows the impact of this modification:

| Before | After |
| --- | --- |
| ![image](https://user-images.githubusercontent.com/583546/219652222-2aca88ad-c4b4-4d6f-9b3f-707df626711e.png) | ![image](https://user-images.githubusercontent.com/583546/219652291-3c1f3246-4ac4-49f8-a087-50949f21b48b.png) |

- The time it takes as measured by XDebug profiler: 55ms (before) vs 27ms (after).
- The number of times the `_build_block_template_result_from_file` is called: 22 (before) vs 5 (after).

## How to test

- Start the bundled environment.
- Set `.env` to production by applying the following patch:

```diff
diff --git a/.env b/.env
index 63a8169f64..9b1d52ac50 100644
--- a/.env
+++ b/.env
@@ -54,10 +54,10 @@ LOCAL_DB_TYPE=mysql
 LOCAL_DB_VERSION=5.7
 
 # The debug settings to add to `wp-config.php`.
-LOCAL_WP_DEBUG=true
-LOCAL_WP_DEBUG_LOG=true
-LOCAL_WP_DEBUG_DISPLAY=true
-LOCAL_SCRIPT_DEBUG=true
+LOCAL_WP_DEBUG=false
+LOCAL_WP_DEBUG_LOG=false
+LOCAL_WP_DEBUG_DISPLAY=false
+LOCAL_SCRIPT_DEBUG=false
 LOCAL_WP_ENVIRONMENT_TYPE=local
 
 # The URL to use when running e2e tests.
```
- Run 100 test of the homepage: `seq 100 | xargs -Iz curl -o /dev/null -H 'Cache-Control: no-cache' -s -w "%{time_starttransfer}\n" http://localhost:8889 | xclip -selection clipboard`.
- Copy the results to a spreadsheet like [this one](https://docs.google.com/spreadsheets/d/1K2u5oOh7rd23HBOz3pPAcAi9P8290E0IV2NZhk_E9i4/edit#gid=0) to measure the changes.

## Notes

I submitted this PR to core instead of Gutenberg because it modifies functions that are no longer there:

- `_get_block_templates_files`: no longer present.
- `get_block_templates`: the corresponding `gutenberg_get_block_templates` is [still present](https://github.com/WordPress/gutenberg/blob/trunk/lib/compat/wordpress-6.1/block-template-utils.php#L53), but it'll be removed once WordPress 6.2 is published. It won't be present in Gutenberg when this PR lands (WordPress 6.3 cycle).