@oandregal 2023-03-10
The [perf job](https://github.com/WordPress/wordpress-develop/actions/runs/4384673981/jobs/7676429393) reports the following (copy/pasted here in a single table, otherwise it's difficult to compare the results):

- Home classic theme

| Metric | This PR | Base | Variance |
| --- | --- | --- | --- |
| wpBeforeTemplate | 24.32 | 25.33 | +3.98% |
|    wpTemplate    |       50.67       |  56.97 | +11.05% |
|     wpTotal      | 75.28  | 80.32 | +6.25% |

- Home block theme

| Metric | This PR | Base | Variance |
| --- | --- | --- | --- |
| wpBeforeTemplate | 28.33  |   31.80  | +10.91% |
|    wpTemplate    | 53.37 | 71.39  | +25.24% |
|     wpTotal      | 81.34 |  102.74  | +20.82% |