@youknowriad 2023-03-21
Can't talk about the performance impact but this looks great to me. Thanks for the updates @oandregal