@felixarntz 2023-04-21
@oandregal How did you get the numbers from the perf CI job? I'm asking since they only really run for `trunk` commits as far as I know. Did you trigger it yourself and look at the workflow run logs? The fact that classic themes see a notable improvement strikes me as potentially incorrect, since the change here should only affect block themes like you're saying.