@felixarntz 2023-04-24
@oandregal 
> Oh. I looked at the [performance CI job that ran for this PR](https://github.com/WordPress/wordpress-develop/actions/runs/4768156334/jobs/8477156436?pr=4097). My understanding was that the numbers under "Print performance test results" was this PR and the numbers under "Print base line test results" was the baseline to compare to. Is that how should I look at them @felixarntz ?

Ah that may be a misunderstanding and slightly confusing. The "base line test results" are not the current stats for `trunk`, but rather a re-run of getting metrics for `trunk` in the same baseline state that was originally measured (which I believe we have set to the last commit hash of WP 6.1.1). So this isn't the state of latest `trunk`.

Looking at the printed "performance test results" is a great idea though. What you need to compare that too though is the similar data for the last commit in `trunk` that is in your branch here.

All of that is of course not how this should work, but to be fair the performance CI workflow in its current stage was not yet meant to be used to assess PRs. Reporting diffs between the last commit in the branch and the PR automatically in a PR comment is high on the todo-list though :)