@oandregal 2023-04-26
This has been committed at https://core.trac.wordpress.org/changeset/55687