@oandregal 2023-07-31
Fix at https://github.com/WordPress/wordpress-develop/pull/4940